Queda la muerte

+++ src/app/academia/tareas/lista-tareas/lista-tareas.component.ts
@@ -6,6 +6,10 @@ import { TareaService } from '../../../services/tarea.service';
 import { TareaEntity } from '../../../interfaces/tarea-entity';
 import { Page } from '../../../interfaces/page';
 import { PaginationComponent } from '../../pagination/pagination/pagination.component';
+import { Usuario } from '../../../interfaces/usuario';
+import { RolUsuario } from '../../../enum/rol-usuario';
+import { AuthService } from '../../../services/auth.service';
+import { fromReadableStreamLike } from 'rxjs/internal/observable/innerFrom';
 
 @Component({
   selector: 'app-lista-tareas',
@@ -18,7 +22,9 @@ export class ListaTareasComponent implements OnInit {
   // Propiedades para almacenar la lista de tareas y la página actual
   tareas: TareaEntity[] = [];
   page: Page<TareaEntity> | null = null;
-
+  //Usuario actual y rol (controlar permisos)
+  usuario: Usuario | null=null;
+  rolUsuario = RolUsuario;
   // Parámetros de paginación y filtrado
   currentPage: number = 0;
   pageSize: number = 10;
@@ -33,31 +39,89 @@ export class ListaTareasComponent implements OnInit {
   loading: boolean = false;
   error: string | null = null;
   activeTab: 'todos' | 'pendientes' | 'vencidas' = 'todos';
-new: any;
+  new: any;
 
-  constructor(private tareaService: TareaService, private router: Router) {}
+  constructor(private tareaService: TareaService, private router: Router, private authService: AuthService) {}
 
   ngOnInit(): void {
-    this.loadTareas();
+    this.authService.currentUser.subscribe(
+      user => {
+        this.usuario=user;
+        if(this.usuario){
+          this.loadTareasSegunRol();
+        }
+      }
+    );
   }
 
-  loadTareas(): void {
+  loadTareasSegunRol(): void{
+    //Sino existe usuario no se puede cargar tareas
+    if(!this.usuario) return;
+
     this.loading = true;
     this.error = null;
     this.isSearchActive = false;
 
-    this.tareaService.getTareas(this.currentPage, this.pageSize, this.sortBy, this.sortDirection).subscribe({
-      next: (page) => {
-        this.page = page;
-        this.tareas = page.content;
+    switch(this.usuario.rol){
+      case RolUsuario.PROFESOR:
+        if(this.usuario.profesorId){
+          this.loadTareasProfesor(this.usuario.profesorId);
+        }
+        break;
+      case RolUsuario.ADMIN:
+        this.loadTodasLasTareas();
+        break;
+      case RolUsuario.ALUMNO:
+        if(this.usuario.alumnoId){
+          this.loadTareasAlumno(this.usuario.alumnoId)
+        }
+        break;
+      default:
+        this.loadTodasLasTareas();
+        break;
+    }
+  }
+
+  loadTareasProfesor(profesorId: number):void{
+    this.tareaService.getTareasByProfesor(profesorId, this.currentPage, this.pageSize, this.sortBy, this.sortDirection).subscribe({
+      next: (page) =>{
+        this.page=page;
+        this.tareas= page.content;
+        this.loading= false;
+      }, error: (err) => {
+        this.error = "Error al cargar las tareas del profesor";
         this.loading = false;
-      },
-      error: (err) => {
-        this.error = 'Error al cargar las tareas. Inténtelo de nuevo más tarde.';
-        this.loading = false;
-        console.error('Error:', err);
+        console.error("Error: ", err);
       }
     });
+  }
+
+  loadTareasAlumno(alumnoId: number):void{
+    this.tareaService.getTareasByAlumno(alumnoId, this.currentPage, this.pageSize, this.sortBy, this.sortDirection).subscribe({
+      next: (page) =>{
+        this.page=page;
+        this.tareas= page.content;
+        this.loading= false;
+      }, error: (err) => {
+        this.error = "Error al cargar las tareas del alumno";
+        this.loading = false;
+        console.error("Error: ", err);
+      }
+    });
+  }
+
+  loadTodasLasTareas(): void{
+    this.tareaService.getTareas(this.currentPage, this.pageSize, this.sortBy, this.sortDirection).subscribe({
+      next: (page)=>{
+        this.page=page;
+        this.tareas=page.content;
+        this.loading=false;
+      }, error: (err)=>{
+        this.error="Error al cargar las tareas";
+        this.loading=false;
+        console.error("Error: ", err);
+      }
+    })
   }
 
   loadPendientes(): void {
@@ -101,8 +165,9 @@ new: any;
   }
 
   search(): void {
+    // Sino existe una búsqueda activa o el filtro está vacío se cargarán las tareas según el rol como estaba previsto
     if (!this.isSearchActive && !this.nombreFilter.trim()) {
-      this.loadTareas();
+      this.loadTareasSegunRol();
       return;
     }
 
@@ -114,7 +179,7 @@ new: any;
     }
 
     this.isSearchActive = true;
-
+    // Si hay algún nombre en el filtro directamente buscamos las tareas
     this.tareaService.searchTareas(this.nombreFilter, this.currentPage, this.pageSize, this.sortBy, this.sortDirection).subscribe({
       next: (page) => {
         this.page = page;
@@ -139,7 +204,7 @@ new: any;
     } else if (this.isSearchActive) {
       this.search();
     } else {
-      this.loadTareas();
+      this.loadTareasSegunRol();
     }
   }
 
@@ -153,7 +218,7 @@ new: any;
       } else if (tab === 'vencidas') {
         this.loadVencidas();
       } else {
-        this.loadTareas();
+        this.loadTareasSegunRol();
       }
     }
   }
@@ -173,30 +238,47 @@ new: any;
     return this.tareaService.tieneDocumento(tarea);
   }
 
-  formatNota(tarea: TareaEntity): string {
-    return this.tareaService.formatNota(tarea);
-  }
-
-  descargarDocumento(id: number): void {
+  descargarDocumento(id: number): void{
     this.tareaService.downloadDocumento(id).subscribe({
-      next: (blob) => {
-        const tarea = this.tareas.find(t => t.id === id);
-        if (tarea && tarea.nombreDocumento) {
-          const url = window.URL.createObjectURL(blob);
-          const a = document.createElement('a');
-          a.href = url;
-          a.download = tarea.nombreDocumento;
-          document.body.appendChild(a);
-          a.click();
-          window.URL.revokeObjectURL(url);
-          a.remove();
+      next: (blob)=>{
+        //Si la llamada tiene éxito obtiene un objeto tipo BLOB
+        const tarea=this.tareas.find(t=> t.id === id);
+        // se comprueba si existe la tarea con el id y si esta contiene un un documento con nombre definido
+        if(tarea && tarea.nombreDocumento){
+            //Se descarga el archivo
+            const url = window.URL.createObjectURL(blob);
+            const a= document.createElement('a');
+            a.href=url;
+            a.download= tarea.nombreDocumento;
+            document.body.appendChild(a);
+            a.click();
+            window.URL.revokeObjectURL(url);
+            a.remove();
         }
-      },
-      error: (err) => {
-        this.error = 'Error al descargar el documento. Inténtelo de nuevo más tarde.';
-        console.error('Error:', err);
+      }, error: (err)=>{
+        this.error= "Error al descargar el documento";
+        console.error("Error: ",err);
       }
     });
+  }
+
+  verEntregas(tareaId:number): void{
+    this.router.navigate(['/tareas', tareaId, 'entregas']);
+  }
+
+  verEntregasProfesor(profesorId: number): void {
+    this.router.navigate(['/entregas'], {
+      queryParams: { profesorId: profesorId, filtro: 'pendientes' }
+    });
+  }
+
+  // Navegación general
+  verTodasLasEntregas(): void {
+    this.router.navigate(['/entregas']);
+  }
+
+  hacerEntrega(tareaId: number): void{
+     this.router.navigate([`/tareas/${tareaId}/entrega`])
   }
 
   verTarea(id: number): void {
@@ -205,35 +287,54 @@ new: any;
   }
 
   editarTarea(id: number): void {
-    const queryParams = { modo: 'edit' };
-    this.router.navigate(['/tareas', id], { queryParams });
+
+    if(this.usuario?.rol === RolUsuario.ADMIN || this.usuario?.rol === RolUsuario.PROFESOR){
+
+      const queryParams = { modo: 'edit' };
+      this.router.navigate(['/tareas', id], { queryParams });
+
+    }
   }
 
   eliminarTarea(id: number): void {
-    if (confirm('¿Está seguro de que desea eliminar esta tarea?')) {
-      this.tareaService.deleteTarea(id).subscribe({
-        next: () => {
-          if (this.activeTab === 'pendientes') {
-            this.loadPendientes();
-          } else if (this.activeTab === 'vencidas') {
-            this.loadVencidas();
-          } else if (this.isSearchActive) {
-            this.search();
-          } else {
-            this.loadTareas();
+    if(this.usuario?.rol === RolUsuario.ADMIN || this.usuario?.rol === RolUsuario.PROFESOR){
+      if (confirm('¿Está seguro de que desea eliminar esta tarea?')) {
+        this.tareaService.deleteTarea(id).subscribe({
+          next: () => {
+
+              this.loadTareasSegunRol();
+
+          },
+          error: (err) => {
+            this.error = 'Error al eliminar la tarea. Inténtelo de nuevo más tarde.';
+            console.error('Error:', err);
           }
-        },
-        error: (err) => {
-          this.error = 'Error al eliminar la tarea. Inténtelo de nuevo más tarde.';
-          console.error('Error:', err);
-        }
-      });
+        });
+      }
     }
   }
 
   nuevaTarea(): void {
-    const queryParams = { modo: 'crear' };
-    this.router.navigate(['/tareas/nuevo'], { queryParams });
+    if(this.usuario?.rol === RolUsuario.ADMIN || this.usuario?.rol === RolUsuario.PROFESOR){
+      const queryParams = { modo: 'crear' };
+      this.router.navigate(['/tareas/nuevo'], { queryParams });
+    }
+  }
+
+  puedeCrearTarea(): boolean {
+    return this.usuario?.rol === RolUsuario.PROFESOR || this.usuario?.rol === RolUsuario.ADMIN;
+  }
+
+   puedeEditarTarea(): boolean {
+    return this.usuario?.rol === RolUsuario.PROFESOR || this.usuario?.rol === RolUsuario.ADMIN;
+  }
+
+   esProfesor(): boolean {
+    return this.usuario?.rol === RolUsuario.PROFESOR;
+  }
+
+   esAlumno(): boolean {
+    return this.usuario?.rol === RolUsuario.ALUMNO;
   }
 
   isTareaVencida(tarea: TareaEntity | null | undefined): boolean {
